requirements.txt
pycozo[embedded]
requests>=2.31.0
pandas
flask
pymupdf
PyYAML
